Move group writing to dedicated module
Separate the group writing mechanism, which is needed by WATO and the
CME snapshot code, from cmk.gui.watolib.groups to eliminate 2 cycles.
