+++ cmk/gui/watolib/groups.py
@@ -6,10 +6,9 @@
 import copy
 import re
 from collections.abc import Sequence
-from typing import Any, Dict, get_args, List, Literal, Optional, Tuple, Type, Union
+from typing import Any, List, Literal, Optional, Tuple, Type, Union
 
 import cmk.utils.paths
-import cmk.utils.store as store
 import cmk.utils.version as cmk_version
 from cmk.utils.type_defs import EventRule, timeperiod_spec_alias
 
@@ -32,12 +31,13 @@ from cmk.gui.htmllib.html import html
 from cmk.gui.http import request
 from cmk.gui.i18n import _
 from cmk.gui.logged_in import user
-from cmk.gui.plugins.watolib.utils import config_variable_registry, wato_fileheader
+from cmk.gui.plugins.watolib.utils import config_variable_registry
 from cmk.gui.utils.html import HTML
 from cmk.gui.utils.urls import makeuri_contextless
 from cmk.gui.valuespec import DualListChoice
 from cmk.gui.watolib.changes import add_change
 from cmk.gui.watolib.global_settings import GlobalSettings, load_configuration_settings
+from cmk.gui.watolib.group_writer import save_group_information
 from cmk.gui.watolib.host_attributes import (
     ABCHostAttribute,
     HostAttributeTopic,
@@ -46,14 +46,10 @@ from cmk.gui.watolib.host_attributes import (
 from cmk.gui.watolib.hosts_and_folders import CREFolder, Folder, folder_preserving_link
 from cmk.gui.watolib.notifications import load_notification_rules, load_user_notification_rules
 from cmk.gui.watolib.rulesets import AllRulesets
-from cmk.gui.watolib.utils import convert_cgroups_from_tuple, format_config_value
+from cmk.gui.watolib.utils import convert_cgroups_from_tuple
 
 if cmk_version.is_managed_edition():
     import cmk.gui.cme.managed as managed  # pylint: disable=no-name-in-module
-
-
-def _clear_group_information_request_cache() -> None:
-    load_group_information.cache_clear()  # type: ignore[attr-defined]
 
 
 def add_group(name: GroupName, group_type: GroupType, extra_info: GroupSpec) -> None:
@@ -196,68 +192,6 @@ def _set_group(
 
     if group_type == "contact":
         hooks.call("contactgroups-saved", all_groups)
-
-
-def save_group_information(
-    all_groups: AllGroupSpecs,
-    custom_default_config_dir: Optional[str] = None,
-) -> None:
-    if custom_default_config_dir:
-        check_mk_config_dir = "%s/conf.d/wato" % custom_default_config_dir
-        multisite_config_dir = "%s/multisite.d/wato" % custom_default_config_dir
-    else:
-        check_mk_config_dir = "%s/conf.d/wato" % cmk.utils.paths.default_config_dir
-        multisite_config_dir = "%s/multisite.d/wato" % cmk.utils.paths.default_config_dir
-
-    _save_cmk_base_groups(all_groups, check_mk_config_dir)
-    _save_gui_groups(all_groups, multisite_config_dir)
-
-    _clear_group_information_request_cache()
-
-
-def _save_cmk_base_groups(all_groups: AllGroupSpecs, config_dir: str) -> None:
-    check_mk_groups: Dict[GroupType, Dict[GroupName, str]] = {}
-    for group_type, groups in all_groups.items():
-        check_mk_groups[group_type] = {}
-        for gid, group in groups.items():
-            check_mk_groups[group_type][gid] = group["alias"]
-
-    # Save Checkmk world related parts
-    store.makedirs(config_dir)
-    output = wato_fileheader()
-    for group_type in get_args(GroupType):
-        if check_mk_groups.get(group_type):
-            output += "if type(define_%sgroups) != dict:\n    define_%sgroups = {}\n" % (
-                group_type,
-                group_type,
-            )
-            output += "define_%sgroups.update(%s)\n\n" % (
-                group_type,
-                format_config_value(check_mk_groups[group_type]),
-            )
-    store.save_text_to_file("%s/groups.mk" % config_dir, output)
-
-
-def _save_gui_groups(all_groups: AllGroupSpecs, config_dir: str) -> None:
-    multisite_groups: Dict[GroupType, Dict[GroupName, GroupSpec]] = {}
-
-    for group_type, groups in all_groups.items():
-        for gid, group in groups.items():
-            for attr, value in group.items():
-                if attr != "alias":  # Saved in cmk_base
-                    multisite_groups.setdefault(group_type, {})
-                    multisite_groups[group_type].setdefault(gid, {})
-                    multisite_groups[group_type][gid][attr] = value
-
-    store.makedirs(config_dir)
-    output = wato_fileheader()
-    for what in get_args(GroupType):
-        if multisite_groups.get(what):
-            output += "multisite_%sgroups = \\\n%s\n\n" % (
-                what,
-                format_config_value(multisite_groups[what]),
-            )
-    store.save_text_to_file("%s/groups.mk" % config_dir, output)
 
 
 def find_usages_of_group(name: GroupName, group_type: GroupType) -> List[Tuple[str, str]]:
