+++ cmk/gui/watolib/sample_config.py
@@ -15,12 +15,13 @@ from cmk.utils.paths import site_cert_file
 from cmk.utils.tags import sample_tag_config, TagConfig
 from cmk.utils.type_defs import EventRule
 
+from cmk.gui.groups import AllGroupSpecs, GroupName
 from cmk.gui.log import logger
 from cmk.gui.plugins.watolib.utils import sample_config_generator_registry, SampleConfigGenerator
 from cmk.gui.userdb import create_cmk_automation_user
 from cmk.gui.watolib.config_domains import ConfigDomainCACertificates
 from cmk.gui.watolib.global_settings import save_global_settings
-from cmk.gui.watolib.groups import AllGroupSpecs, GroupName, save_group_information
+from cmk.gui.watolib.group_writer import save_group_information
 from cmk.gui.watolib.hosts_and_folders import Folder
 from cmk.gui.watolib.notifications import save_notification_rules
 from cmk.gui.watolib.rulesets import FolderRulesets
